refactor(cli): consolidate implementation into package and fix CI test failures
